+++ doc/version_history.rst
@@ -4,6 +4,12 @@
 Version History
 ===============
 
+v0.9.3
+------
+
+* Updated for the latest ts-tcpip and ESS CSC XML.
+* Added commands and telemetry for the DIMM XML.
+
 v0.9.2
 ------